fix(buildimage): Deprecate the old buildimages update task
