+++ tasks/buildimages.py
@@ -4,7 +4,8 @@ import os
 
 from invoke import Context, task
 
-from tasks.pipeline import update_circleci_config, update_gitlab_config, update_test_infra_def
+from tasks.libs.ciproviders.circleci import update_circleci_config
+from tasks.libs.ciproviders.gitlab_api import update_gitlab_config, update_test_infra_def
 
 
 @task(
